Extract section icon into trait
