+++ packages/forms/src/Components/Section.php
@@ -4,6 +4,7 @@ namespace Filament\Forms\Components;
 
 use Closure;
 use Filament\Support\Concerns\HasExtraAlpineAttributes;
+use Filament\Support\Concerns\HasIcon;
 use Illuminate\Contracts\Support\Htmlable;
 use Illuminate\Support\Str;
 
@@ -13,6 +14,7 @@ class Section extends Component implements Contracts\CanConcealComponents, Contr
     use Concerns\CanBeCompacted;
     use Concerns\EntanglesStateWithSingularRelationship;
     use HasExtraAlpineAttributes;
+    use HasIcon;
 
     /**
      * @var view-string
@@ -24,8 +26,6 @@ class Section extends Component implements Contracts\CanConcealComponents, Contr
     protected string | Htmlable | Closure $heading;
 
     protected bool | Closure | null $isAside = null;
-
-    protected string | Closure | null $icon = null;
 
     protected bool | Closure $isFormBefore = false;
 
@@ -70,13 +70,6 @@ class Section extends Component implements Contracts\CanConcealComponents, Contr
         return $this;
     }
 
-    public function icon(string | Closure | null $icon): static
-    {
-        $this->icon = $icon;
-
-        return $this;
-    }
-
     public function getDescription(): string | Htmlable | null
     {
         return $this->evaluate($this->description);
@@ -110,11 +103,6 @@ class Section extends Component implements Contracts\CanConcealComponents, Contr
     public function isAside(): bool
     {
         return (bool) ($this->evaluate($this->isAside) ?? false);
-    }
-
-    public function getIcon(): ?string
-    {
-        return $this->evaluate($this->icon);
     }
 
     public function formBefore(bool | Closure $condition = true): static
