+++ packages/support/src/Concerns/HasIcon.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Filament\Support\Concerns;
+
+use Closure;
+
+trait HasIcon
+{
+    protected string | Closure | null $icon = null;
+
+    public function icon(string | Closure | null $icon): static
+    {
+        $this->icon = $icon;
+
+        return $this;
+    }
+
+    public function getIcon(): ?string
+    {
+        return $this->evaluate($this->icon);
+    }
+}
